+++ kinto/core/storage/postgresql/migrations/migration_015_016.sql
@@ -1,0 +1,4 @@
+DROP FUNCTION collection_timestamp(uid VARCHAR, resource VARCHAR);
+
+-- Bump storage schema version.
+INSERT INTO metadata (name, value) VALUES ('storage_schema_version', '16');